When a new gift is added to the cart, update the locally stored cart.
The productData and itemConfig are not all of the item data that would normally be in cartData, some logic happens in cartService to create this object. Since I did not want to duplicate that logic, I opted to clear the local cartData and recalculate the fees when loading the cart instead.