Reword branded checkout device id instructions
Previously, the instructions made it sound like a new merchant account was needed. It's not. In most cases, Cru's main merchant account should be reused for the new device id.

This came out of a conversation I had with Steve Hackney regarding Story Runner's desire for branded checkout on their site.